+++ controllers/register.js
@@ -1,0 +1,33 @@
+const handleRegister = (req, res, db, bcrypt) => {
+  const {email, name, password} = req.body
+
+  const hash = bcrypt.hashSync(password, 10);
+
+  db.transaction(trx => {
+    trx.insert({
+      hash,
+      email
+    })
+    .into('login')
+    .returning('email')
+    .then(loginEmail => {
+      return trx('users')
+        .returning('*')
+        .insert({
+          email,
+          name,
+          joined: new Date()
+        })
+        .then(user => {
+          res.json(user[0])
+        })     
+    })
+    .then(trx.commit)
+    .catch(trx.rollback)
+  })
+    .catch(err => res.status(400).json('Unable to Register User'))
+}
+
+module.exports = {
+  handleRegister
+}
